Created new scenes, created algorithms for win conditions.

+++ Assets/Scripts/Core/GameData.cs
@@ -5,7 +5,7 @@ namespace TicTacToe
     [CreateAssetMenu(fileName = "Game Data", menuName = "Core/Game Data")]
     public class GameData : ScriptableObject
     {
-        [Min(3)] public float BoardWidth = 3;
-        [Min(3)] public float BoardHeight = 3;
+        [Min(3)] public int BoardWidth = 3;
+        [Min(3)] public int BoardHeight = 3;
     }
 }
